Fix linting issues for pre-existing tests
